+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v 0.0.2
+
+- added Helper.real_timestamp_of_id
+- added default configuration so app doesn’t crash if user didn’t set config
+- added Erlang Node name as an option for nodes
+- updated documentation
+
 ## v 0.0.1
 
 Initial Release